requirements.txt
nornir
nornir_napalm
nornir_netmiko
nornir_utils
pandas
pyyaml